General improvements and bug fixes (#43)
* Feature: Inline statements

Implements cherriesandmochi/gdmaim#8.
Combines lines of code into one wherever it can and retains scopes.

* Fix inline statements merging empty lines and comments

* Enhance: Implement inline scoping

Shrinks the output code even more.
If it finds that the statements in a scope have reduced to a single
line, it will convert that indented line into an inline one.
Plus additional checks for exceptions.

* Refactor tokenization

Rename 'identation' -> 'indentation'.
Fix Enum values to support bitmasking.
Separate annotations to its own type.

* Fix own-line annotations in inline statements

Annotations that annotate the next line instead of the current one no
longer are separated via a semicolon.

* Fix multiline annotation parameters

* Fix export_subgroup not ending in newline

* Add more annotations to syntax highlighting

* Fixes cherriesandmochi/gdmaim#40

* Fixes cherriesandmochi/gdmaim#42

* Refactor statement break parsing

* Fix inline statements not merging statement broken lines

* Implement parsing getter setters for variables

* Fix merge error leftover code

* Fix support lambas in variable declaration defaults

* Clean up some AST Node logging outputs

* Fix 4.4 Array type error

Implements cherriesandmochi/gdmaim#45

* Support class-scoped anonymous Enums

* Fix support function blocks within parenthesis

Implements recursive inlining, which avoids issues caused by lambda functions in lists or as a callback argument

* Fix: Anonymous Enum keys error in a classless script

* Fix: export_category needs to be followed by a newline

* Remove debug prints

* Fix is_indentation typo

* Comply with warning_ignore decorator rules

warning_ignore will be followed by a space
warning_ignore_start and warning_ignore_restore will be followed by a newline

* Better Inline Statements lambda detection

* Optimise: do not put redundant semicolons or spaces

* Optimise: do not put redundant space from statement breaks

* Fix: Inline Statements lambda function argument list support

* Add missing Variants to syntax highlighter list

* Fix: inline lambdas now detected

Previously lambdas would only detect if they were indented
This new detection can spot a lambda anywhere in the line

* Fix: scopes with warning_ignore should not collapse

* Fix: Remove leftover indentations caused by Dicts

Dictionaries can have its key and value split between different lines,
thus making element pairs take up multiple lines and confuse
the indentation level, which lambdas depend on. This fixes it

* Fix: Unexpected commas should end recursive level

Commas detected outside of the recursive call's brackets are to be
treated as if the comma separates a higher level's recursive code's
elements

* Refactor: "identation" -> "indentation"

* Update Readme

* Fix: nested IFs and ELSEs inlining rules

The line with the ELSE statement actually is the ELSE to the
last IF statement in a previous line, not the first one.
This change respects that by making sure some IF scopes do not collapse

* Fix: String tokens missing decos in str()

* Fix: support first line funcs and lists

* Fix: exiting lambdas didn't dec temp scopes

* Fix: support ELIF as ELSE

* Fix: nested inline IF detection

* Fix: Recursive collapse warning_ignore

---------

Co-authored-by: Kaleb Reid <78945904+Kaleb-Reid@users.noreply.github.com>